Implement 'top_k' in preprocessing.py.
