+++ requirements.txt
@@ -1,5 +1,7 @@
-tensorflow==2.2.0
-scikit-learn==0.23.0
-numpy==1.18.4
-pandas==1.0.5
-matplotlib==3.3.2
+tensorflow~=2.0.0
+scikit-learn~=0.22.1
+numpy~=1.18.1
+pandas~=1.0.1
+matplotlib~=3.1.3
+nmslib~=2.0.6
+setuptools~=46.0.0
