+++ sui/dl/__init__.py
@@ -1,8 +1,6 @@
 """sui.dl
 Deep learning algorithm implementations
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-model_class_list = ["sui.dl.nn.PNN() => Product-based Neural Networks"]
+from __future__ import absolute_import, division, print_function, unicode_literals
+from .pnn import PNN
+from .gru4rec import GRU4Rec
