+++ sui/ml/__init__.py
@@ -1,8 +1,5 @@
 """sui.ml
 Machine learning algorithm implementations
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-model_class_list = ["sui.ml.mf.FunkSVD() => FunkSVD"]
+from __future__ import absolute_import, division, print_function, unicode_literals
+from .mf import FunkSVD, BiasSVD, SVDpp, BPR, ALS
